+++ batch/src/main/java/com/yachaerang/batch/controller/BatchJobController.java
@@ -3,8 +3,8 @@ package com.yachaerang.batch.controller;
 import com.yachaerang.batch.domain.dto.JobStartResponseDto;
 import com.yachaerang.batch.domain.dto.JobStatusResponseDto;
 import com.yachaerang.batch.exception.GeneralException;
-import com.yachaerang.batch.service.BatchJobService;
-import com.yachaerang.batch.service.JobStatusService;
+import com.yachaerang.batch.service.job.BatchJobService;
+import com.yachaerang.batch.service.job.JobStatusService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.batch.core.JobExecutionException;
 import org.springframework.format.annotation.DateTimeFormat;
