+++ batch/src/main/java/com/yachaerang/batch/service/job/JobStatusService.java
@@ -1,4 +1,4 @@
-package com.yachaerang.batch.service;
+package com.yachaerang.batch.service.job;
 
 import com.yachaerang.batch.domain.dto.JobStatusResponseDto;
 import com.yachaerang.batch.exception.GeneralException;
